fix: value service
