+++ packages/core/src/index.ts
@@ -1,6 +1,7 @@
 export * from './utils';
 export * from './definitions';
 export * from './entity';
+export * from './value-service';
 export * from './communication';
 export * from './enums';
 export * from './fetchers';
