+++ packages/core/src/value-service/index.ts
@@ -1,0 +1,1 @@
+export { ValueService } from './ValueService';
